+++ pkg/metricsservice/server.go
@@ -107,10 +107,9 @@ func (s *GrpcServer) GetRawMetricsStream(request *api.RawMetricsRequest, stream 
 					return err
 				}
 			}
-		case val, open := <-doneCh:
-			if open && !val {
-				logger.V(10).Info(fmt.Sprintf("Channel closed for subscriber %s", request.GetSubscriber()))
-			}
+		case <-doneCh:
+			logger.V(10).Info("Raw metrics stream terminated after subscriber unsubscribed", "subscriber", request.GetSubscriber())
+			return nil
 		}
 	}
 }